backend_api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
geoalchemy2[shapely]==0.14.4
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
